app: move logic from onRender and onEvent to onLoop
Avoid having logic on the render and event handling functions, and keep
it as much as possible in the loop function.

Use the event function to update player actions, etc. but in a way that
the computed logic upon those updates is run on the loop function.

Use the render loop solely for rendering.

These updates required some extra if's. I try to keep them by checking
in the order:

1. Is a PVP or versus AI
2. Whose turn it is

The way code is written may not be the most compact of optimal, but I
aim for clarity.