app: text: fix memory leak misuse of existing texture
An existing texture pointer was being assigned with
SDL_CreateTextureFromSurface over and over in the rendering loop. This
led to a memory leak.

As a fix, the texture is created and freed inside the text function,
instead of being a global object.

Performance is not taken into consideration for this fix. It may not be
the most efficient way to implement this.